Backend/requirements.txt
Flask==2.3.2
psutil==5.9.4
Flask-CORS==5.0.0
gunicorn==20.1.0
python-dotenv==0.21.1
